boreco/requirements.txt
asgiref==3.5.0
azure-core==1.23.1
azure-storage-blob==12.11.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0
django-cors-headers==3.11.0
django-filter==21.1
django-import-export==2.8.0
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-yasg==1.20.0
et-xmlfile==1.1.0
flake8==4.0.1
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
isodate==0.6.1
itypes==1.2.0
Jinja2==3.1.1
MarkupPy==1.14
MarkupSafe==2.1.1
mccabe==0.6.1
msrest==0.6.21
oauthlib==3.2.0
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.3
Pillow==9.1.0
psycopg2-binary==2.9.1
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.3.0
pymongo==4.1.1
pyparsing==3.0.8
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
xlrd==2.0.1
xlwt==1.3.0

